requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6
itsdangerous==2.1.2